app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
transformers==4.44.2
torch==2.4.1
PyPDF2==3.0.1
python-docx==1.1.2
sqlite3
requests==2.32.3